api/pyproject.toml
[tool.poetry]
name = "halver-api"
version = "0.1.0"
description = ""
authors = ["Inioluwa Akinyosoye <inisoye@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.4"
environs = { extras = ["django"], version = "^9.5.0" }
psycopg2-binary = "^2.9.5"
pillow = "^9.3.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
djangorestframework-camel-case = "^1.3.0"
aiohttp = "^3.8.3"
celery = "^5.2.7"
redis = "^4.4.0"
drf-spectacular = "^0.25.1"
dj-rest-auth = { extras = ["with-social"], version = "^2.2.7" }
django-ipware = "^4.0.2"
flower = "^1.2.0"
django-import-export = "^3.0.2"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.0.2"}

[tool.poetry.group.dev.dependencies]
django-stubs = { extras = ["compatible-mypy"], version = "^1.13.1" }
types-requests = "^2.28.11.7"
django-debug-toolbar = "^3.8.1"
docformatter = "^1.5.1"

[tool.poe.tasks]
dev = "poetry run python manage.py runserver"
test-m = "poetry run python manage.py makemigrations --dry-run"
make-m = "poetry run python manage.py makemigrations"
show-m = "poetry run python manage.py showmigrations"
m = "poetry run python manage.py migrate"
create-su = "poetry run python manage.py createsuperuser"
collectstatic = "poetry run python manage.py collectstatic"
celery = "poetry run celery -A core worker -l info"
flower = "poetry run celery -A core flower"
celery-purge = "poetry run celery -A core purge"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
